+++ src/utils/obj/index.ts
@@ -1,4 +1,6 @@
 export { default as deepEqualsObj } from './deepEqualsObj'
+export * from './deepMergeObj'
 export { default as deleteProp } from './deleteProp'
 export { default as equalsObj } from './equalsObj'
+export { default as mergeObj } from './mergeObj'
 export { default as toObjString } from './toObjString'