+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## [1.2.0] 04-26-2025
+
+### Added
+
+- `keysOf` and `valuesOf` to reduce bundle size when using `Object.keys` or `Object.values` multiple times.
+- `mergeObj` and `deepMergeObj` for merging objects.
+- `DeepMergeResult<T>` and `ShallowMergeResult<T>` to type `mergeObj` and `deepMergeObj` strictly.
+
+### Changed
+
+- `isObj` now returns `false` for arrays to correctly distinguish between arrays and objects.
+- `deepArrEquals` handles objects deeply as well now.
+
 ## [1.1.0] 03-16-2025
 
 - Add built-in `Math` properties to minify